Trying to get a simple scenario to compile
I have found that all the files in the module end up in the same ns-3 module folder, leading to some very interesting issues. First, I have had to modify some file names to ensure that I don't end up overwriting any files. Secondly, I have had to change the includes on a lot of files so that creating a scenario does not make the system explode.

I have yet been unable to compile an example. However the files themselves seem to be compiling find. My issues continue to be related to the template classes which have issues with the logging system used by ns-3.

I shall go to sleep for today.